+++ admin/settings.php
@@ -75,6 +75,9 @@ if(isset($_POST['submitted'])) {
 	if(isset($_POST['sitename'])) { 
 		$SITENAME = htmlentities($_POST['sitename'], ENT_QUOTES, 'UTF-8'); 
 	}
+	if(isset($_POST['sitedesc'])) { 
+		$SITEDESC = htmlentities($_POST['sitedesc'], ENT_QUOTES, 'UTF-8'); 
+	}
 	if(isset($_POST['siteurl'])) { 
 		$SITEURL = tsl($_POST['siteurl']); 
 	}
@@ -151,6 +154,8 @@ if(isset($_POST['submitted'])) {
 		$xmls = new SimpleXMLExtended('<?xml version="1.0" encoding="UTF-8"?><item></item>');
 		$note = $xmls->addChild('SITENAME');
 		$note->addCData($SITENAME);
+		$note = $xmls->addChild('SITEDESC');
+		$note->addCData($SITEDESC);
 		$note = $xmls->addChild('SITEURL');
 		$note->addCData($SITEURL);
 		$note = $xmls->addChild('TEMPLATE');
@@ -218,7 +223,9 @@ get_template('header', cl($SITENAME).' &raquo; '.i18n_r('GENERAL_SETTINGS'));
 			<?php	if ( $fullpath != $SITEURL ) {	echo '<p style="margin:-15px 0 20px 0;color:#D94136;font-size:11px;" >'.i18n_r('LABEL_SUGGESTION').': &nbsp; <code>'.$fullpath.'</code></p>';	}	?>
 		</div>
 		<div class="clear"></div>
-		
+		<div class="widesec">
+			<p><label for="sitedesc"><?php i18n('LABEL_WEBSITEDESC'); ?>:</label><textarea class="text" id="sitedesc" name="sitedesc"><?php echo $SITEDESC; ?></textarea></p>
+		</div>
 		<p class="inline" ><input name="prettyurls" id="prettyurls" type="checkbox" value="1" <?php echo $prettychck; ?>  /> &nbsp;<label for="prettyurls" ><?php i18n('USE_FANCY_URLS');?></label></p>
 				
 		<div class="leftsec">